% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmApp-package.R
\docType{package}
\name{sdmApp-package}
\alias{sdmApp-package}
\title{sdmApp: A User-Friendly Application for Species Distribution Modeling}
\description{
sdmApp provides a 'shiny' graphical user interface that lets non-expert
R users build species distribution models through a reproducible
workflow: uploading raster predictors and species occurrence data,
inspecting predictor correlation, selecting predictors with
\pkg{CENFA}, applying spatial block cross-validation with \pkg{blockCV},
fitting a range of models, and exporting results together with the
underlying R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Abson-dev/sdmApp}
  \item \url{https://abson-dev.github.io/sdmApp/}
  \item Report bugs at \url{https://github.com/Abson-dev/sdmApp/issues}
}

}
\author{
Aboubacar HEMA
}
\keyword{internal}
