% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmApp.R
\name{sdmApp}
\alias{sdmApp}
\title{Using the interactive GUI - sdmApp}
\usage{
sdmApp()
}
\value{
An interactive GUI
}
\description{
Using the interactive GUI - sdmApp
}
\examples{

#load the package
library(sdmApp)
sdmApp()
}
