% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmApp_PA.R
\name{sdmApp_PA}
\alias{sdmApp_PA}
\title{Plot presence/absence map}
\usage{
sdmApp_PA(x)
}
\arguments{
\item{x}{\code{Raster object}}
}
\value{
a \code{ggplot object}
}
\description{
Plot presence/absence map
}
\examples{
r <- raster::raster(system.file("extdata","AETI.tif",package = "sdmApp"))
r <- r > 4000
sdmApp_PA(r)
}
