% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmApp_RasterPlot.R
\name{sdmApp_RasterPlot}
\alias{sdmApp_RasterPlot}
\title{Plot a  raster}
\usage{
sdmApp_RasterPlot(x)
}
\arguments{
\item{x}{\code{Raster object}}
}
\value{
a \code{ggplot object}
}
\description{
Plot a  raster
}
\examples{
r <- raster::raster(system.file("extdata","AETI.tif",package = "sdmApp"))
sdmApp_RasterPlot(r)
}
